+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * `ItemWrapper#findWrapped` method to obtain wrapper of specific type
 * Item default `toString` implementation
 * `AdaptViewGroup#init` convenience method that takes `ChangeHandler` directly
+* `LineNumberId` utility to get current line number in source code
 * A collection of wrappers:
   * `BackgroundWrapper`
   * `FrameWrapper`